Update patch set 7
Patch Set 7:

Hm. Thought of another alternative to holding a naked pointer.
As implemented with DTLSTransport, the webrtc:DTLSTransport holds ownership to the cricket::DTLStransport, which holds ownership to the cricket::IceTransport.

What if we made the IceTransport object (presently with a pointer) hold a scoped_refptr to the webrtc::DTLStransport, which in turn holds on tightly to the cricket::DTLSTransport, which holds on tightly to the cricket::IceTransport?

That way, we have a chain of documented ownership all the way down; we just have to make webrtc::IceTransport do the right thing when Clear() has been called on the webrtc::DTLSTransport.

Patch-set: 7
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0