Update patch set 8
Patch Set 8:

(7 comments)

I was unable to use #include in place of forward declarations because of the dependency rules. @mirko will have to comment about whether the p2p target should be a dependency of the libwebrtc-api target or not. I am still not certain what the purpose of the libwebrtc-api target is.

Patch-set: 8
Reviewer: Gerrit User 5634 <5634@58829da1-049c-39fb-b951-ebdcd0984767>