Update patch set 8
Patch Set 8:

(1 comment)

> Patch Set 8:
> 
> (7 comments)
> 
> I was unable to use #include in place of forward declarations because of the dependency rules. @mirko will have to comment about whether the p2p target should be a dependency of the libwebrtc-api target or not. I am still not certain what the purpose of the libwebrtc-api target is.

I like how you reworked this CL, thanks for doing this!

With regards to forward declarations, I talked with kwiberg@ yesterday and yes, there are situations where they are useful ans this is one of them (until we will have C++ modules, it seems a good trade-off to avoid types leakage).

So I would say that it is OK to forward declare cricket::IceTransportInternal and cricket::PortAllocator (but it seems possible to remove the forward declaration of webrtc::IceTransportWithPointer in pc/dtls_transport.h).

The purpose of libjingle_peerconnection_api is unclear, we are going to try to formalize something during the next week (together with the guideline about .h/.cc and usage of interfaces).

Patch-set: 8