Update patch set 1
Patch Set 1:

Do we need the simulated time for this? I imagine that it's unfeasible to use simulated time for PeerConnection tests?

My suggestion is to remove the TimeController parts from this CL. Use a TaskQueue directly instead. For repeated activities, use RepeatingTaskHandle. This will keep this CL smaller and AFICT it should only affect a small part of the code.

We can revisit the TimeController later if/when we want to port scenario tests to this.

(trying to minimize the work for now)

Patch-set: 1
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>