Update patch set 1
Patch Set 1:

(3 comments)

If we remove the time controller (for now) and fix the minor things mentioned below I think this is ready. Just go trough all comments and check the grammar, there are some errors still. Also, please replace the URLs with something more guaranteed to be persistent, such as a reference to a standard, the title of the document etc. That way future readers can web search for the source even if it has moved.

Patch-set: 1
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>