Update patch set 2
Patch Set 2:

(3 comments)

My main concern at this stage is I don't know much of the behavior of the TaskRunner from the upstream. I would suggest a file scope comment in the upstream CL to detail the behavior of the TaskRunner, e.g, if it has the thread affinity or is just a sequencer (please see SingleThreadTaskRunner vs SequencedTaskRunner in Chromium), if a task is cancellable and how we manage the captures.

Patch-set: 2
Reviewer: Gerrit User 6646 <6646@58829da1-049c-39fb-b951-ebdcd0984767>