Update patch set 10
Patch Set 10:

(4 comments)

Added the 2x probe as discussed offline. Having that probe makes a lot of sense, as we might allocate up to 2x the max media bitrate to the BitrateObserver in BitrateAllocator. And then we might as well probe at that level too. With this change, this CL should purely lead to fewer probes, but the probes will still be at the same level.

Patch-set: 10
Reviewer: Gerrit User 5023 <5023@58829da1-049c-39fb-b951-ebdcd0984767>