Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review+1
> 
> I still think the rule makes sense in principle, for the same reason it makes sense to use ArrayView instead of const vector&. But since raw pointers isn't a type we control, it's not possible to have the same convenient implicit conversions that we get with ArrayView.

I think I agree. For conversion in one direction, it's a pity there's no .get, analogous to unique_ptr .get. And a way to construct or assign an optional from a const T* would make sense too. But we're not in control of the optional type. 

I the examples, caller and/or callee have already chosen optional as the underlying type, and if that's unlikely to change, I don't oppose optional strongly.

I think I'm more reluctant to encourage const ref optional as a return type than as an argument type. I'm thinking of refactorings like

Before:

Foo* foo() { return foo_ ? &*foo_ : nullptr; }

After:

Foo* foo() { return other_ ? &other_->foo : nullptr; }

Where the interface offers access to a foo that may or not be present, and where ownership of the storage is moved from an optional to some more complex and optional member (I'm thinking of other_ as a unique_ptr). To to the same refactoring with an unchanged const ref optional return type, one would need to duplicate the value in a long-lived optional. 

In contrast, when constructing an optional to pass as a const ref argument to the function, it needs to live only over the duration of the call.

Passing and returning optionals by value avoids life time complications.

Patch-set: 4