Update patch set 5
Patch Set 5:

(1 comment)

> Patch Set 5:
> 
> I wonder if we can get rid of some of the layers by having accessors.
> For example, if the end goal is to set the latency of a particular audio source, you might be able to separate the logic for finding that audio source from actual plumbing.
> 
> For example:
> 
> If "track" is a remote track, you might want to find its receiver, and use the receiver to find the audio source. Pseudo-code:
> 
> for each receiver in peer_connection {
>   if (receiver.track == track) {
>     receiver->source->or->whatever->SetLatency()
>   }
> }
> 
> This avoids adding getters, setters and callbacks to every single layer in the stack, even though it is only the audio source has a notion of "SetLatency()", the rest becomes simple mapping helper. Would that work?

The problem that actual connection to the streams is in rtp_receiver. If you open it, you will see that it is where remote track and source are created in pair. And most of other plumbing happens because we need to reach NetEq from audio_receive_stream.h

Patch-set: 5
Reviewer: Gerrit User 8286 <8286@58829da1-049c-39fb-b951-ebdcd0984767>