Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > > Patch Set 3:
> > > > 
> > > > Why couldn't the receiving side always send feedback on request? We don't expect to get a request unless it's negotiated, so we could just always accept it and send feedback, worst case, we'll send feedback even though we shouldn't have done?
> > > 
> > > The current behavior is to send feedback periodically. If feedback on request is negotiated we want to turn this off, that's the main reason why we need this information.
> > 
> > Don't we want periodic feedback still? If we lose a lot of packets with the feedback request bit set it seems like we still want to have feedback sent eventually?
> 
> I think it's less complex to just have the sender request feedback until it has been received.
> The other difference between periodic and requested is that with the requested feedback, we clear away any data prior the requested, while the periodic would not. It just becomes more complex to reason about the logic if once in a while we get unsolicited periodic feedback with slightly different semantics.

Ok, so you don't see any risks with large amount of loss or delayed deliveries on the uplink resulting in no feedback being sent? Not a common case, but what we build should be robust to all cases.

I think we may still have logic in the BWE that reacts to no feedback being received for a long time, would that be impacted by this, or perhaps that was only looking at RTCP RR?

Patch-set: 3
Reviewer: Gerrit User 5525 <5525@58829da1-049c-39fb-b951-ebdcd0984767>