Update patch set 4
Patch Set 4:

> Possibly, since media/base/ is a legacy API directory. Do you suspect that this change might break someone, or is that unlikely?

Main usecase for cricket::VideoCapturer was to act as a local video source for PeerConnection.

Application code could create subclasses of cricket::VideoCapturer, but since the previous cl was landed, there's no way to pass a subclass like that directly to PeerConnection{Factory}. It does implement the VideoSourceInterface though, so it could be used for things that expect a VideoSourceInterface rather than the VideoTrackSourceInterface passed to PeerConnection. I think that's not so likely.

Patch-set: 4
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0