Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> Before I start diving into the code... Is there some document or similar that explains at a high level what you're trying to accomplish? The commit message makes it sound like you're building something just like task queues, but a separate implementation with separate interfaces, and doesn't explain why we'd want that.

The idea is to allow injecting the implementation and allowing usage of a simulated time implementation. I considered if it was possible to use the TaskQueue interface system, but that is not finished yet and doesn't seem to provide the right level of abstraction (for instance it doesn't provide any explicit interface for repeated tasks and requires a boolean return value indicating ownership, making the abstraction a bit low level)

I don't know the timeline of the TaskQueue refactor, maybe when that's done, it could be changed to support a similar pattern, making it usable instead? For now I needed something to provide the functionality for the change in Scenario tests to allow it to share more code with the network emulation.

Patch-set: 7
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>