Update patch set 7
Patch Set 7:

Is there a ddoc?

I'm concerned with the TQ infrastructure getting bloated. We're moving towards a virtual interface, there's already a TQ for testing class + another class that's essentially technical debt now, which is a single threaded TQ-like class (but not actually a TQ) that's still in use in many places and I'm not aware of any work going on to clean that up.

Patch-set: 7
CC: Gerrit User 5508 <5508@58829da1-049c-39fb-b951-ebdcd0984767>