Update patch set 4
Patch Set 4:

(5 comments)

> Patch Set 4:
> 
> (5 comments)
> 
> Do you think it is important to merge updated regions on frame drops? Otherwise, you could consider if it makes things simpler to include the timestamp of the previous frame with the UpdateRect. Then each encoder can cache the timestamp of the previous frame, and ignore the UpdateRect if timestamps don't match.
>

I believe it's important because it will help to speed-up encoding of unchanged regions. The idea with timestamps is sound, but we are rewriting timestamps in VideoStreamEncoder, so it's not so straightforward. Some unique counter may be added instead or Id may be reused. But I see merit in merging update_rects and don't find the logic to be too complicated.
 
> For screenshare, ideally we shouldn't drop frames in the VideoAdapter, but adjust how often we sample the screen.

This is an action item for the future. But it's a separate issue from the frames dropped in VideoStreamEncoder due to mediaOpt or busy encoder queue.

Patch-set: 4