Update patch set 1
Patch Set 1:

Use a const pointer instead, to avoid requiring a change of signature in case we'll need to read values at runtime in the future. Keep an instance of the default config as a member and fallback to that one if nullptr is provided, the default implementation is really cheap so it should add any significant overhead. (this is similar to how it's done for GoogCC)

Patch-set: 1
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>