Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> > Patch Set 1:
> > 
> > Use a const pointer instead, to avoid requiring a change of signature in case we'll need to read values at runtime in the future. Keep an instance of the default config as a member and fallback to that one if nullptr is provided, the default implementation is really cheap so it should add any significant overhead. (this is similar to how it's done for GoogCC)
> 
> There is also opposite opinion on this subject: "implement something only when you actually need it"
> https://testing.googleblog.com/2017/08/code-health-eliminate-yagni-smells.html"

Since this is it isn't really more complex to make it future proof I don't think there's a problem of violating yagni here. As it's intended as an API it will be annoying to change the interface in the future.

It might be worthwhile considering if we should always only allow checking trials on construction, but for now the pattern is that we allow checking trials at run time.

Patch-set: 1
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>