Remove VCMEncodedFrameCallback and VCMGenericEncoder
This CL removes generic encoder and moves some of the logic to
VideoStreamEncoder. Things related to timing frames are extracted to
a dedicated class called FrameEncodeTimer, which is fully owned by
VideoStreamEncoder. It is made thread safe since we can't guarantee
that encode start and encode done happens on the same thread, and we
can't safely post and wait for answer on encoder worker thread.

Bug: webrtc:10164
Change-Id: I9b26f734473b659e4093c84c09fb0ed441290e40