clang-tidy helper script, with clang static analyzer included.
This CL makes it easy to invoke the tool for single cc files,
without having to install or configure anything.
It boils down to:
% tools_webrtc/clang_tidy.py path/to/file.cc [clang-tidy-args...]
If any issue is encountered, a colored diagnostic will be printed.
By default, it also includes checks from clang analyzer.

Warning! This is linux-only, and doesn't use an old version of
clang-tidy based on LLVM 5.0. USE WITH CARE.

bug: webrtc:8793
Change-Id: I8964f2b939408326cc349c5f0ac0dfcff2da24c5