Add ::Connect method to the media transport interface
In order to enable ::Connect method, we also need to split the factory and create a method that creates media transport, but doesn't connect it.

So far media transport was connecting right away after creation. We would however want to expose some of the settings in SDP. SDP is created before connection is connected (and before ICE transport is created), and so we would like to be able to get the settings from the caller to the callee.

Bug: webrtc:9719
Change-Id: I1dc2f30c9a2dae8b3db04f14c8b334cd1b3ab5ab