Update patch set 3
Patch Set 3:

(1 comment)

I have updated this change to pass a decryption state back to this line that does the checking. I am not 100% sure if the CritScope is required here. The ManagedFrame is called on the network_tc but I am  not sure which thread the video_receive_stream is called on.

Patch-set: 3
Reviewer: Gerrit User 6930 <6930@58829da1-049c-39fb-b951-ebdcd0984767>