Update patch set 1
Patch Set 1:

I think this is an improvement, but still not so nice to have the send streams borrow the transport's queue. If for some reason that is essential, maybe the transport should provide a PostTask method instead? What are the blockers for letting the streams own their own task queues? 

With MediaTransport, it get's a bit strange to have a task queue for borrowing, the model is rather that we call methods on the transport objects to pass encoded frames, possibly directly from various encoder threads, and then the transport will internally post work to its queue.

Patch-set: 1
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>