Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I think this is an improvement, but still not so nice to have the send streams borrow the transport's queue. If for some reason that is essential, maybe the transport should provide a PostTask method instead? What are the blockers for letting the streams own their own task queues? 
> 
> With MediaTransport, it get's a bit strange to have a task queue for borrowing, the model is rather that we call methods on the transport objects to pass encoded frames, possibly directly from various encoder threads, and then the transport will internally post work to its queue.

The reason we need to share the TaskQueue atm is to mainly to ensure synchronization on creation/start/stop/destruction of send streams. It is also used for transport related parts of VideoSendStreamImpl that generally should be moved into the RTP video sender.

The latter should resolve itself naturally when we finish the move of those parts. The former is trickier and we don't really have a final design for how we should manage synchronisation of such things, and in particular how the BitrateAllocator should interact with transport, call and send streams respectively.

If this is landed, we should be able to make channel_send to use its own encoder queue. If we to that we should be able to provide a PostTask on the transport controller.

Patch-set: 1
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>