Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Looks good and like a scalable approach, but wouldn't the API be easier with a simple argument to the PlatformThread ctor? What's the need for ScopedRegisterThreadForDebugging?

This way, it also works for threads not created with PlatformThread. But perhaps we don't care about being able to register those?

Also, having a single-purpose class to own the registration and clean it up in its destructor feels like a good way to organize this code. But I could still do that, by adding a ScopedRegisterThreadForDebugging member to PlatformThread.

Patch-set: 6
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>