Update patch set 5
Patch Set 5:

(8 comments)

Thanks for the review!
found a possible bug in casting char to unsigned char and bit manipulations.
ArrayView is causing some problems, mostly in unit tests because of the reinterpret_casts but i guess this is a necessity because of lack of byte object.
the missing DCHECK might get added in the commit for supporting enum classes. depending on which one lands first.

Patch-set: 5
Reviewer: Gerrit User 8244 <8244@58829da1-049c-39fb-b951-ebdcd0984767>