Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 1:
> 
> (2 comments)
> 
> > Naming: Just "ScopedEventSync" should be sufficient---that it's using thread-local storage is an implementation detail, and it's possible to parse the current name as meaning "thread scoped" (I did), which is confusing.
> 
> I'd prefer ScopedYieldSomething. Maybe just ScopedYield, or ScopedYieldHook, ScopedYieldPolicy, ...

"ScopedYieldPolicy" is the same thing I was about to suggest, but didn't since I wasn't sure it would be enough of an improvement. But with two people independently suggesting it, it feels like a more solid recommendation.

Also, ping on that unit test? It doesn't have to be big---just install a mock yielder that expects to be called once, and wait for an event. For bonus points, use two threads with independent yielders whose execution is interleaved in such a way that the test would fail unless the yield policy was really thread local.

Patch-set: 2
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>