Adds RefPtr class for simple use cases of scoped_refptr.
The system for scoped_refptr and RefCountInterface is quite flexible and
supports advanced use cases such as implementing copy on write etc.

This is useful in some cases, but sometimes it's overkill. This CL
introduces the RefPtr class which allows using scoped_refptr with any
preexisting class without further customization.

Bug: webrtc:10365
Change-Id: Ia7f4f670fa19af30d8770705cbe0093d50649952