Update patch set 4
Patch Set 4:

(1 comment)

> See the bug conversation for more details.

I've had another look at the bug, and I still don't get it. According to https://bugs.chromium.org/p/webrtc/issues/detail?id=10321#c4, there's no explicit signalling of the sender's ntp timestamps over the wire; they are estimated and then used to populate sender_ntp_time_ms_. What I don't understand is why we can't populate ntp_time_ms_ instead? Why is it useful with two different timestamps?

I totally agree that it's useful with as good an estimate as we can get of the sender's wallclock time when each frame was captured. The question is, why two?

And if we want to use some kind of "system local" time for rendering and synchronization, that ought to go into timestamp_us_. (It's not clear to me if and how that field is populated on the receive side).

Patch-set: 4
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>