Update patch set 4
Patch Set 4:

(2 comments)

It seems fine to me but I'm missing the same information on high level audio API, as I was doing in the second version of my patch[1]. I guess this was the problem about not breaking prominent API users like chrome? Is there a procedure to modify the API for AudioSinkInterface? It would be useful to modify the OnData method to pass a struct instead of actual parameters, to allow to add more fields in future if necessary. Or just reusing AudioFrame, although it has low level rtp timestamp.

Also, please observe comments on "audio/channel_receive.cc" as estimation is done after the sink callbacks.

[1] https://bugs.chromium.org/p/webrtc/issues/detail?id=10321#c11

Patch-set: 4
CC: Gerrit User 8941 <8941@58829da1-049c-39fb-b951-ebdcd0984767>