Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > 3. There is no real difference, I would prefer the ifdef because if keeps everything in the realm of the preprocessor/compiler instead of relying on GN's flexibility.
> > 
> > That's still going to be depend on the flag from GN, but makes the code a lot less readable.
> 
> This is true if we look at the code in isolation, but the overall context is as bad as having the #ifdef because we would have 2 .cc files to define the same function and which one ends up in the binary is decided by GN.

Yes, that would be just as bad.

My suggestion is to add a separate DegradedCallFactory in a separate target, then inject that using PeerConnectionDependencies in Chromium. This is similar to how we do with other large components such as encoders etc.

Patch-set: 4
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>