Update patch set 1
Patch Set 1:

(3 comments)

Excellent! I've been thinking on and off for a few years now that someone ought to do this...

Standardizing on the name ThreadChecker is probably less than ideal. SequencedTaskChecker would be better (since a thread is a specific type of sequence) but the "Task" in that name is misplaced for threads. How about "SequenceChecker"?---that seems to be what Chromium calls it.

(And if you end up having to rename the file anyway, consider moving it to a subdirectory, since rtc_base/ is an API directory.)

Patch-set: 1
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>