Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> This concern basically means that public_deps should never used, if I understand it correctly.

We have tried to eliminate it completely (it causes some trouble for alien build systems used downstream). Current state is that I think we allow it to easy transitions when moving targets or source files around, but it's otherwise discouraged in webrtc.

> Since we do use it all over,

Not really, if we're talking about webrtc. I count to 35 uses of public_deps in webrtc BUILD.gn files, compared to 900+ plain deps.

So if we can solve the problem with one more public_deps, I don't think that's a big deal, but if we need to add half a dozen, I don't like it so much.

Patch-set: 3
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>