Update patch set 2
Patch Set 2:

(1 comment)

> The previous attempt failed because somewhere in the process, correct operation depended on calling the send functions on the datachannel with id=-1 and queue empty.
> 
Yikes. Out of curiosity, what exactly is dependent on this? It would be nice if there were a test that tested this behavior and stopped us from landing breaking changes like this.

Change looks good, I'll wait for a test to give it an lgtm.

Patch-set: 2
Reviewer: Gerrit User 5800 <5800@58829da1-049c-39fb-b951-ebdcd0984767>