Update patch set 8
Patch Set 8: Code-Review+1

> PTAL again. :)
That looks great, this is more symmetrical now!

Please add a comment in the description about htons and co being not guaranteed to do the right thing (with MSDN link), because this is an additional change on top of "extra hton functions".

Patch-set: 8
Reviewer: Gerrit User 7332 <7332@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1